ml/pyproject.toml
[project]
name = "evoke-ml"
version = "0.1.0"
description = "ML service for Evoke - audio to visual embedding"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.35.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "numpy>=1.24.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pymilvus>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
